+++ src/index.js
@@ -1,5 +1,6 @@
 import * as Models from './models/index.js'
+import * as utils from './utils/index.js'
 
 import { generateSSORequestParams, tokenPayloadVerifier, messageToSign } from './sso.js'
 
-export default { Models, generateSSORequestParams, tokenPayloadVerifier, messageToSign }
+export default { Models, utils, generateSSORequestParams, tokenPayloadVerifier, messageToSign }